(MAINT) Restructure `dsc-lib-jsonschema`
This change refactors the `dsc-lib-jsonschema` library without modifying
any behavior.

This change:

- Splits the functions in the `transforms` module out into submodules
  and re-exports them from `transforms` - this keeps referencing the
  functions the way it was before but makes it easier to navigate the
  files, given their length.
- Makes the unit tests for `schema_utility_extensions` mirror the
  structure from the source code.
- Makes the integration tests for `transform` mirror the structure from
  the source code.
