+++ lib/dsc-lib-jsonschema/src/tests/schema_utility_extensions.rs
