--- lib/dsc-lib-jsonschema/src/tests/transforms/mod.rs
@@ -1,4 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-//! Unit tests for [`dsc-lib-jsonschema::transforms`]
