+++ lib/dsc-lib-jsonschema/tests/integration/transforms/idiomaticize_externally_tagged_enum.rs
