+++ lib/dsc-lib-jsonschema/tests/integration/transforms/idiomaticize_string_enum.rs
