--- lib/dsc-lib-jsonschema/tests/integration/transforms/idiomaticizing/enums/mod.rs
@@ -1,7 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-//! Integration tests for idiomaticizing the generated schemas for `enum` items.
-
-#[cfg(test)] mod string_variants;
-#[cfg(test)] mod externally_tagged;
