--- lib/dsc-lib-jsonschema/tests/integration/transforms/idiomaticizing/mod.rs
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-//! Integration tests for idiomaticizing the generated schemas. The schemas that [`schemars`]
-//! generates are sometimes non-idiomatic, especially when you use annotation keywords for variants
-//! and fields.
-
-#[cfg(test)] mod enums;
